playing with a pattern where the node drives everything

+++ src/constants.zig
@@ -35,8 +35,8 @@ pub const verify: bool = true;
 /// Maximum size of the message pool
 pub const default_worker_message_pool_capacity: u32 = 10_000;
 
-pub const connection_outbox_capacity: u32 = 5_000;
-pub const connection_inbox_capacity: u32 = 5_000;
+pub const connection_outbox_capacity: usize = 5_000;
+pub const connection_inbox_capacity: usize = 5_000;
 
 pub const topic_max_queue_capacity: usize = 10_000;
 pub const publisher_max_queue_capacity: usize = 5_000;
